+++ src/ui/radio-group/Option.tsx
@@ -1,18 +1,9 @@
 import { useRef } from 'react';
-import { OptionType } from 'src/constants/articleProps';
+import { OptionProps } from 'src/constants/articleProps';
 import { Text } from 'src/ui/text';
 import { useEnterSubmit } from './hooks/useEnterSubmit';
 
 import styles from './RadioGroup.module.scss';
-
-type OptionProps = {
-	value: OptionType['value'];
-	title: OptionType['title'];
-	selected: OptionType;
-	groupName: string;
-	onChange?: (option: OptionType) => void;
-	option: OptionType;
-};
 
 export const Option = (props: OptionProps) => {
 	const { value, title, selected, groupName, onChange, option } = props;
